% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.seurat.R
\name{cell.filter.seurat}
\alias{cell.filter.seurat}
\title{cell.filter.seurat}
\usage{
cell.filter.seurat(seurat, sample.name, metrics_summary, more_nFeature_RNA = 200, Less_nFeature_RNA = 8000, percent.mt=10)
}
\arguments{
\item{seurat}{Seurat object}

\item{sample.name}{single cell RNA sequening sample name}

\item{metrics_summary}{summary metrics}

\item{more_nFeature_RNA}{High cutoffs for filtering cells that have unique feature counts (default is 200)}

\item{Less_nFeature_RNA}{low cutoffs for filtering cells that have unique feature counts (default is 2500)}

\item{percent.mt}{low cutoffs for filtering cells that have >n percent mitochondrial counts (default is 5)}
}
\value{
a Seurat object containing only the relevant subset of cells
}
\description{
A wrapper function to cell.filter.seurat
}
\details{
Creates a Seurat object containing only a subset of the cells in the original object.
}
