% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.cell.typing.seurat.R
\name{cell.typing.seurat}
\alias{cell.typing.seurat}
\title{cell.typing.seurat}
\usage{
cell.typing.seurat(
  seurat,
  marker = "Puram.2017.HNSCC",
  cell.typing.method = c("NTP", "ES", "Average"),
  level = c("cell", "cluster"),
  wd,
  slot = c("scale.data", "count.data", "data"),
  assay = "RNA",
  ntp.dir,
  rda.dir,
  NTP.g.filter.method = c("sd", "mad", "none"),
  NTP.gene.filter.cutoff = 0.3,
  NTP.distance = c("cosine", "correlation"),
  NTP.norm.method = c("none", "row.std"),
  mc.cores = 1
)
}
\arguments{
\item{seurat}{Seurat object}

\item{marker}{Cell markers to use in cell typing, character or List (identifier or StudyName or User defined gene list)}

\item{cell.typing.method}{cell typing method, c("NTP", "ES", "Average"), (default = "NTP")}

\item{wd}{working directory}

\item{slot}{assay data type of seurat object, c("scale.data", "count.data", "data")}

\item{assay}{Assay of seurat object}

\item{ntp.dir}{Output directory of NTP}

\item{rda.dir}{Path of the RData saving directory}

\item{NTP.g.filter.method}{Method of gene filtering in NTP c(sd (Default), mad, none)}

\item{NTP.gene.filter.cutoff}{Cut-off score of standard deviation in NTP}

\item{NTP.distance}{Method of calculating distance in NTP, either c("correlation" or "cosine").}

\item{NTP.norm.method}{Method of normalization in NTP, either c("none", "row.std")}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\description{
Wrapper function for cell typing of seurat object
}
