% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.cellranger.R
\name{CellrangerCount}
\alias{CellrangerCount}
\title{CellrangerCount}
\usage{
CellrangerCount(cellranger.path, fastq.dir, cellranger.ref.dir, output.dir, sample.name, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{cellranger.path}{Cell Ranger program path}

\item{fastq.dir}{FastQC output directory}

\item{cellranger.ref.dir}{Directory of Cell Ranger reference file}

\item{output.dir}{Output directory}

\item{sample.name}{sample name}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
feature-barcode matrices and Secondary analysis (e.g., dimensionality reduction, cell clustering, and differential expression)
}
\description{
A wrapper function to run CellrangerCount
}
\details{
CellrangerCount takes FASTQ files from fastQC and performs alignment, filtering, barcode counting, and UMI counting.
}
\references{
Massively parallel digital transcriptional profiling of single cells. GXY Zheng. (2017).
}
\seealso{
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/using/count}
}
