% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09.viz.R
\name{cnv.distribution}
\alias{cnv.distribution}
\title{cnv.distribution}
\usage{
cnv.distribution(seurat, wd, marker="Puram.2017.HNSCC.TME", slot=c("scale.data", "count.data", "data"))
}
\arguments{
\item{seurat}{seurat object}

\item{wd}{working directory}

\item{marker}{cell type marker}

\item{slot}{assay data type of seurat object, c("scale.data", "count.data", "data")}
}
\description{
Visualize result cnv distribution of scTyper
}
