% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.inferCNV.R
\name{cts.geneSetCluster}
\alias{cts.geneSetCluster}
\title{cts.geneSetCluster}
\usage{
cts.geneSetCluster(cset, rda.dir, perm.t.resList, fc=0.05, bp = 1000000)
}
\arguments{
\item{cset}{cnv Set}

\item{rda.dir}{rData directory}

\item{perm.t.resList}{list of t-test result}

\item{bp}{base pair}

\item{fc}{fold change}
}
\value{
cell type specific geneClustList
}
\description{
A wrapper function to cts.geneSetCluster
}
\details{
get gene cluster
}
