% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.function.R
\name{df2gr}
\alias{df2gr}
\title{df2gr}
\usage{
df2gr(df, seqnames, start, end, strand)
}
\arguments{
\item{df}{dataframe}

\item{seqnames}{A character vector of recognized names for the column in df that contains the chromosome name (sequence name) associated with each genomic range.}

\item{start}{A character vector of recognized names for the column in df that contains the start positions of the genomic ranges.}

\item{end}{A character vector of recognized names for the column in df that contains the end positions of the genomic ranges.}

\item{strand}{A character vector of recognized names for the column in df that contains the strand associated with each genomic range.}
}
\value{
GRanges object
}
\description{
A wrapper function to make dataframe to GRange
}
\details{
the workhorse behind the coercion method from data.frame to GRanges.
}
