% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09.viz.R
\name{draw.heatmap}
\alias{draw.heatmap}
\title{draw.heatmap}
\usage{
draw.heatmap(
  seurat,
  wd,
  run.inferCNV = TRUE,
  slot = c("scale.data", "count.data", "data"),
  marker = "Puram.2017.HNSCC.TME"
)
}
\arguments{
\item{seurat}{seurat object}

\item{wd}{working directory}

\item{run.inferCNV}{whether run inferCNV (default=TRUE)}

\item{slot}{assay data type of seurat object, c("scale.data", "count.data", "data")}

\item{marker}{cell type marker}
}
\description{
Visualize heatmap of scTyper
}
