% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.fastqc.R
\name{fastqc}
\alias{fastqc}
\title{fastqc}
\usage{
fastqc(
  fastqc.path,
  fastq.dir,
  sample.name,
  fq1.idx = "_R1_001.fastq",
  fq2.idx = "_R2_001.fastq",
  output.dir,
  run.cmd = TRUE,
  mc.cores = 1
)
}
\arguments{
\item{fastqc.path}{FastQC program path}

\item{fastq.dir}{FastQC output directory}

\item{sample.name}{sample name}

\item{fq1.idx}{Index of the FASTQ file (Read 1)}

\item{fq2.idx}{Index of the FASTQ file (Read 2)}

\item{output.dir}{Output directory}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Quality check report for sequence data. (e.g., .html)
}
\description{
A wrapper function to run fastQC
}
\details{
FastQC aims to provide a QC report that detects problems originating from either the sequencer or the starting library material.
}
\references{
FastQC: a quality control tool for high throughput sequence data. Andrews S. (2010).
}
\seealso{
\url{http://www.bioinformatics.babraham.ac.uk/projects/fastqc}
}
