% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.fastqc.R
\name{fastqc.summary}
\alias{fastqc.summary}
\title{fastqc.summary}
\usage{
fastqc.summary(fastqc.df, fq1.idx="_R1_001.fastq", fq2.idx="_R2_001.fastq")
}
\arguments{
\item{fastqc.df}{fastqc output directory}

\item{fq1.idx}{Index of the FASTQ file (Read 1)}

\item{fq2.idx}{Index of the FASTQ file (Read 2)}
}
\value{
fastqc dataframe
}
\description{
A wrapper function to make fastqc dataframe
}
\details{
make table of fastqc using qc outputs.
}
