% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.fastqc.R
\name{fastqc.table}
\alias{fastqc.table}
\title{fastqc.table}
\usage{
fastqc.table(qc.dir)
}
\arguments{
\item{qc.dir}{fastqc output directory}
}
\value{
fastqc dataframe
}
\description{
A wrapper function to make fastqc dataframe
}
\details{
make table of fastqc using qc outputs.
}
