% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.inferCNV.R
\name{fil.infercnv_obj}
\alias{fil.infercnv_obj}
\title{fil.infercnv_obj}
\usage{
fil.infercnv_obj(infercnv_obj, cset, rda.dir, gprofiler.resList, cell.type.set, cts.geneClustList)
}
\arguments{
\item{infercnv_obj}{infercnv object}

\item{cset}{cnv Set}

\item{rda.dir}{rData directory}

\item{gprofiler.resList}{gprofiler result List}

\item{cell.type.set}{cell type set}

\item{cts.geneClustList}{cell type specific geneClustList}
}
\value{
infercnv object
}
\description{
A wrapper function to fil.infercnv_obj
}
\details{
get gene cluster
}
