% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.inferCNV.R
\name{get.geneClust}
\alias{get.geneClust}
\title{get.geneClust}
\usage{
get.geneClust(gr, bp=1000000)
}
\arguments{
\item{gr}{GRanges object}

\item{bp}{base pair}
}
\value{
gene clust List
}
\description{
A wrapper function to get.geneClust
}
\details{
get gene cluster
}
