% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.function.R
\name{get.markerList}
\alias{get.markerList}
\title{get.markerList}
\usage{
get.markerList(marker)
}
\arguments{
\item{marker}{Signature_list or Signature name of scTyper db or User-defined list of marker genes}
}
\value{
marker list
}
\description{
get markerList from scTyper databse
}
