% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.fastqc.R
\name{get.qc.report}
\alias{get.qc.report}
\title{get.qc.report}
\usage{
get.qc.report(qc.dir)
}
\arguments{
\item{qc.dir}{qc directory}
}
\description{
QC Reports
}
\details{
Provides FASTQC report that summarizes the QC processing steps
}
