% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.function.R
\name{GSEA.EnrichmentScore2}
\alias{GSEA.EnrichmentScore2}
\title{GSEA.EnrichmentScore2}
\usage{
GSEA.EnrichmentScore2(gene.list, gene.set, weighted.score.type = 1, correl.vector = NULL)
}
\arguments{
\item{gene.list}{gene signature list}

\item{gene.set}{gene set}

\item{weighted.score.type}{Type of weighted score}

\item{correl.vector}{correlation vector}
}
\description{
Run GSEA(Gene Set Enrichment Analysis)
}
