% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.inferCNV.R
\name{infercnv2cset}
\alias{infercnv2cset}
\title{infercnv2cset}
\usage{
infercnv2cset(infercnv_obj, pdata)
}
\arguments{
\item{infercnv_obj}{an infercnv object}

\item{pdata}{phenotype data}
}
\value{
cset
}
\description{
A wrapper function to infercnv2cset
}
\details{
Creation of cset using infercnv object.
}
