% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.function.R
\name{invalid}
\alias{invalid}
\title{invalids}
\usage{
invalid(x)
}
\arguments{
\item{x}{value to be tested}
}
\value{
Bool
}
\description{
Test if a value is missing, empty, or contains only NA or NULL values.
}
