% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.function.R
\name{make.color.set}
\alias{make.color.set}
\title{make.color.sample}
\usage{
make.color.sample(n)
}
\arguments{
\item{n}{the number  of color sample, integer}

\item{col.label}{color label, character}
}
\description{
make ramdom colors
}
