% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.makeSet.R
\name{make.eset}
\alias{make.eset}
\title{make.eset}
\usage{
make.eset(expr, pdata=NULL, fdata=NULL, verbose=TRUE)
}
\arguments{
\item{expr}{expression data}

\item{pdata}{phenotype data}

\item{fdata}{feature data}

\item{verbose}{default TRUE, Class to writing verbose messages to a connection or file.}
}
\value{
expression set
}
\description{
For the expression data are transformed to a file with extension .eSet
}
\details{
store expression data in ExpressionSet format for convenient analysis
}
