% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.makeSet.R
\name{make.seurat.eset}
\alias{make.seurat.eset}
\title{make.seurat.eset}
\usage{
make.seurat.eset(seurat, slot=c("scale.data", "count.data", "count.data"), fdata, output.dir=NULL, save=T)
}
\arguments{
\item{seurat}{Seurat object}

\item{slot}{assay data type of seurat object, c("scale.data", "count.data", "data")}

\item{fdata}{feature data}

\item{output.dir}{output directory}

\item{save}{whether save}
}
\value{
expression set
}
\description{
A wrapper function to make.seurat.eset
}
\details{
make expression set using seurat
}
