% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.makeSet.R
\name{make.seurat.fdata}
\alias{make.seurat.fdata}
\title{make.seurat.fdata}
\usage{
make.seurat.fdata(seurat, gene.ref.gtf, rda.dir)
}
\arguments{
\item{seurat}{Seurat object}

\item{gene.ref.gtf}{gene reference gtf file}

\item{rda.dir}{rData directory}
}
\value{
feature data
}
\description{
A wrapper function to make.seurat.fdata
}
\details{
make feature data using seurat
}
