% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.makeSet.R
\name{make.seurat.scale.eset}
\alias{make.seurat.scale.eset}
\title{make.seurat.scale.eset}
\usage{
make.seurat.scale.eset(seurat, fdata, output.dir)
}
\arguments{
\item{seurat}{Seurat object}

\item{fdata}{feature data}

\item{output.dir}{output directory}
}
\value{
expression set
}
\description{
A wrapper function to make.seurat.scale.eset
}
\details{
make expression set using the scaled seurat
}
