% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.cellranger.R
\name{make.stat_summary}
\alias{make.stat_summary}
\title{make.stat_summary}
\usage{
make.stat_summary(count.dir, sample.name, output.dir, pheno.df)
}
\arguments{
\item{count.dir}{cellragner count ouput directory}

\item{sample.name}{cell sample name}

\item{output.dir}{Output directory}

\item{pheno.df}{phenotype dataframe(reference an instruction manual)}
}
\value{
data summary file
}
\description{
A wrapper function to run make.stat_summary
}
\details{
make data summary file
}
