% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.inferCNV.R
\name{perm.subcset.t}
\alias{perm.subcset.t}
\title{perm.subcset.t}
\usage{
perm.subcset.t(cset, cell.type.set, rda.dir, ordered= FALSE, levels= c(1,0), mc.cores=5)
}
\arguments{
\item{cset}{cnv Set}

\item{cell.type.set}{cell type set}

\item{rda.dir}{rData directory}

\item{ordered}{order bool, default FALSE. Sort descending vs. ascending}

\item{levels}{confidence level of the interval.}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
list of t-test result
}
\description{
A wrapper function to perm.subcset.t
}
\details{
Performs t-tests on cnv subset
}
