% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.function.R
\name{perm.ttest}
\alias{perm.ttest}
\title{perm.ttest}
\usage{
perm.ttest(eset, g.st,  level=NULL, t.test=F, permp=T, permp.exact=NULL, ordered=T, mc.cores=1,...)
}
\arguments{
\item{eset}{expression set}

\item{g.st}{group subset}

\item{permp}{bool, defalut TRUE. Calculating permuted T test p-values or not}

\item{ordered}{order bool, default FALSE. Sort descending vs. ascending}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{levels}{confidence level of the interval.}
}
\value{
t test result
}
\description{
Run permutation T-test
}
\details{
Performs sample t-tests on vectors of data.
}
