% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.function.R
\name{preRanked.GSEA}
\alias{preRanked.GSEA}
\title{preRanked.GSEA}
\usage{
preRanked.GSEA(expr, SIGDB, weighted.score.type = 0, correl.vector = NULL, n.cutoff=1,mc.cores=1)
}
\arguments{
\item{expr}{expression metrix}

\item{SIGDB}{gene signature list}

\item{weighted.score.type}{Type of weight score}

\item{correl.vector}{correlation vector}

\item{n.cutoff}{number of cutoff}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\description{
a universal gene set enrichment analysis tools
}
