% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08.report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(envList, qc.dir, output.dir)
}
\arguments{
\item{envList}{R environment list}

\item{qc.dir}{qc directory}

\item{output.dir}{output directory}
}
\value{
pdf file include data processing result information
}
\description{
Reports the result of using scTyper()
}
\details{
Provides a report that summarizes the processing steps and visualized tables and plots for the processed results. The report file is automatically generated recording the workflows of the data processing steps, the options used in the processing, and the outcome results.
}
