% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.inferCNV.R
\name{run.inferCNV}
\alias{run.inferCNV}
\title{run.inferCNV}
\usage{
run.inferCNV(seurat, assay, output.dir, rda.dir, fdata, pheno_info = pheno.df, feature.to.test, cells.test_reference, cells.test_excluded, fc.cutoff = 0.05, cutoff.gene.cluster=0.05, min_mean_expr_cutoff = 0.1, window_length = 101, smooth_ends = TRUE, recenter_method = "median", ordered = FALSE, inv_log = TRUE, sd_amplifier = 1.5, bp = 1000000, sd.cut = 0.3, mc.cores=1)
}
\arguments{
\item{seurat}{Seurat object}

\item{assay}{Name of assay to pull data from seurat object}

\item{output.dir}{output directory}

\item{rda.dir}{rData directory}

\item{fdata}{feature information}

\item{pheno_info}{phenotype information}

\item{feature.to.test}{featuest to test either "tissue.type" or "cell.type"}

\item{cells.test_reference}{a vector containing the classifications of the reference (normal) cells to use for infering cnv}

\item{cells.test_excluded}{cell type to exclude functional enrichment analysis}

\item{fc.cutoff}{fold change cutoff}

\item{cutoff.gene.cluster}{A cutoff P-value for filtering out the gene clusters (calculated from GO analysis)}

\item{min_mean_expr_cutoff}{the minimum mean value allowed for a gene to be retained in the expression matrix.}

\item{window_length}{length of window (number of genes) for the moving average}

\item{smooth_ends}{perform smoothing at the ends of the chromosomes (default:TRUE)}

\item{recenter_method}{method to select the center of the cell expression value. (default:'mean', option:'mean', 'median')}

\item{ordered}{order bool, default FALSE. Sort descending vs. ascending}

\item{inv_log}{mean values will be determined based on (2^x -1)}

\item{sd_amplifier}{multiplicative factor applied to the standard deviation to alter the noise range (default: 1.5)}

\item{bp}{base pair}

\item{sd.cut}{standard deviation cutoff}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Seurat object
}
\description{
A wrapper function to run.inferCNV
}
\details{
CNV inference
}
