% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.seurat.R
\name{run.seurat.process}
\alias{run.seurat.process}
\title{run.seurat.process}
\usage{
run.seurat.process(count.dir = count.dir, rda.dir = rda.dir, project = proj.name, metrics_summary, pheno_info = pheno.df, sample.name = sample.name, min.cells = sum(metrics_summary$Estimated.Number.of.Cells)*0.001, min.features = 200, scale.factor = 10000, vars.to.regress=c("nCount_RNA", "percent.mt"), selection.method = "vst", more_nFeature_RNA = 200, Less_nFeature_RNA = 8000, percent.mt = 10, normalize=TRUE, assay = 'RNA', clust.dims=1:10, resolution=seq(0.6, 2, 0.2), random.seed=1234)
}
\arguments{
\item{count.dir}{cellragner count ouput directory}

\item{rda.dir}{Path of the RData saving directory}

\item{project}{project name}

\item{metrics_summary}{cellranger summary metrics}

\item{sample.name}{Sample name}

\item{percent.min.cells}{Include genes with detected expression in at least this many cells. Will subset the raw.data matrix as well. To reintroduce excluded genes, create a new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many genes are detected.}

\item{scale.factor}{Sets the scale factor for cell-level normalization(10,000 by default)}

\item{vars.to.regress}{Variables to regress out (previously latent.vars in RegressOut). For example, nUMI, or percent.mito.}

\item{selection.method}{How to choose top variable features. Choose one of 'vst', 'mean.var.plot', 'dispersion'}

\item{more_nFeature_RNA}{High cutoffs for filtering cells that have unique feature counts (default is 200)}

\item{Less_nFeature_RNA}{low cutoffs for filtering cells that have unique feature counts (default is 8000)}

\item{percent.mt}{low cutoffs for filtering cells that have >n percent mitochondrial counts (default is 10)}

\item{normalize}{use log normalization}

\item{assay}{Assay to use}

\item{dims}{A vector of the dimensions to use in construction of the SNN grouph.}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities.}

\item{random.seed}{Seed of the random number generator.}
}
\value{
feature-barcode matrices and Secondary analysis (e.g., dimensionality reduction, cell clustering, and differential expression)
}
\description{
A wrapper function to run run.seurat.process
}
\details{
CellrangerCount takes FASTQ files from fastQC and  performs alignment, filtering, barcode counting, and UMI counting.
}
\references{
Massively parallel digital transcriptional profiling of single cells. GXY Zheng. (2017).
}
\seealso{
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/using/count}
}
