% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTyper.R
\name{scTyper}
\alias{scTyper}
\title{scTyper}
\usage{
scTyper(seurat.object, marker, wd, output.name, pheno.fn, qc = FALSE, run.cellranger=FALSE, norm.seurat=FALSE, cell.typing.method, level, run.inferCNV=TRUE, proj.name, fastqc.path, fastq.dir, fq1.idx, fq2.idx, cellranger.path, cellranger.ref.dir=NULL, percent.min.cells=0.1, min.features=200, percent.mt=10, vars.to.regress=c("nCount_RNA", "percent.mt"), dims=1:10, resolution=2, slot=c("scale.data", "count.data", "data"), assay="RNA", NTP.g.filter.method=c("sd", "mad", "none"), NTP.gene.filter.cutoff=0.3, NTP.distance=c("cosine","correlation"), NTP.norm.method, gene.ref.gtf=NULL, feature.to.test = c("tissue.type", "cell.type"), cells.test_excluded="Epithelial", cells.test_reference = "immune", fc.cutoff=0.05, cutoff.gene.cluster=0.05, malignant.cell.type="Epithelial", report.mode=TRUE, mc.cores = 1)
}
\arguments{
\item{seurat.object}{Seurat object, if users have pre-processed seurat object, user have to insert seurat object as input}

\item{marker}{Cell markers to use in cell typing, character or List (identifier or StudyName or User defined gene list)}

\item{wd}{Working directory}

\item{output.name}{Output directory name}

\item{pheno.fn}{Phenotype file path}

\item{qc}{Whether to execute FASTQC (default=FALSE)}

\item{run.cellranger}{whether to excute cellranger count (default=FALSE)}

\item{norm.seurat}{whether to normalize seurat object (default=FALSE)}

\item{cell.typing.method}{cell typing method, c("NTP", "ES", "Average"), (default = "NTP")}

\item{level}{Indicate the cell assignment level (cell or cluster)}

\item{run.inferCNV}{Indicate whether ‘malignant cell typing by inferCNV process run}

\item{proj.name}{Project name}

\item{fastqc.path}{FastQC program path}

\item{fastq.dir}{FastQC output directory}

\item{fq1.idx}{Index of the FASTQ file (Read 1)}

\item{fq2.idx}{Index of the FASTQ file (Read 2)}

\item{cellranger.path}{Cell Ranger program path}

\item{cellranger.ref.dir}{Directory of Cell Ranger reference file}

\item{percent.min.cells}{Cutoff to filter features containing minimum percent of cells}

\item{min.features}{Cutoff to filter cells containing minimum number of features}

\item{percent.mt}{Cutoff for filtering cells that have >n percent mitochondrial counts}

\item{vars.to.regress}{Variables to regress out}

\item{dims}{A vector of the dimensions to use in construction of the SNN graph.}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities.}

\item{slot}{Data type of Seurat object, c("scale.data", "count.data", "data")}

\item{assay}{Assay of Seurat object}

\item{NTP.g.filter.method}{Method to filter genes in NTP}

\item{NTP.gene.filter.cutoff}{Cutoff to filter genes of in NTP}

\item{NTP.distance}{NTP distance method, a character, either c("correlation" or "cosine").}

\item{NTP.norm.method}{NTP normalization method, either c("none", "row.std")}

\item{gene.ref.gtf}{Path of GTF file including genomic location for genes}

\item{feature.to.test}{Column header name of the meta data in Seurat object (select the cell groups for T.test) either "tissue.type" or "cell.type"}

\item{cells.test_excluded}{A value indicates the cells to be excluded in T.test}

\item{cells.test_reference}{A value indicates the cells to use as be excluded in T.test}

\item{fc.cutoff}{Cutoff of fold change}

\item{cutoff.gene.cluster}{A cutoff P-value for filtering out the gene clusters (calculated from GO analysis)}

\item{malignant.cell.type}{Cell type to assign malignant cell}

\item{report.mode}{Generate report file}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Seurat object
}
\description{
Run scTyper
}
