% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.function.R
\name{update.seqnames}
\alias{update.seqnames}
\title{update.seqnames}
\usage{
update.seqnames(seqnames)
}
\arguments{
\item{seqnames}{A character vector of recognized names for the column in df that contains the chromosome name (sequence name) associated with each genomic range.}
}
\value{
sequence name
}
\description{
updata the column in df that contains the chromosome name (sequence name)
}
