% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.update.db.R
\name{update.sig.db}
\alias{update.sig.db}
\title{update.sig.db}
\usage{
update.sigTyper.db(sig.db.path, db.name=c("CellMarker", "sigTyper.db"), output.dir=system.file("/data",package = "scTyper") )
}
\arguments{
\item{sig.db.path}{Path of sig.db.txt}

\item{output.dir}{storage path of sig.db marker}
}
\description{
Update sig.db
}
